rename path for page components
