+++ components/[pageId]/BoardView/BoardView.tsx
