+++ components/[pageId]/BoardView/index.ts
