+++ components/[pageId]/DocumentView/components/PageBanner.tsx
