+++ components/common/BoardEditor/focalboard/src/components/viewTitle.tsx
@@ -2,7 +2,7 @@
 // See LICENSE.txt for license information.
 import ImageIcon from '@mui/icons-material/Image'
 import CharmEditor, { ICharmEditorOutput } from 'components/common/CharmEditor/CharmEditor'
-import { PageCoverGalleryImageGroups } from 'components/pageId/DocumentView/components/PageBanner'
+import { PageCoverGalleryImageGroups } from 'components/[pageId]/DocumentView/components/PageBanner'
 import { randomIntFromInterval } from 'lib/utilities/random'
 import { Page, PageContent } from 'models'
 import React, { useCallback, useState } from 'react'
